added upgrade handler test
